fix dependencies and includes
